requirements.txt
aiofiles==0.6.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5; python_version >= "3.5"
chardet==4.0.0
click==7.1.2
fastapi==0.63.0
google-api-core==1.26.2
google-auth==1.28.0
google-cloud-automl==2.2.0
google-cloud-core==1.6.0
google-cloud-storage==1.37.0
google-cloud-vision==2.2.0
google-crc32c==1.1.2; python_version >= "3.5"
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpcio==1.36.1
h11==0.12.0
idna==2.10
libcst==0.3.17
mypy-extensions==0.4.3
packaging==20.9
proto-plus==1.18.1
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20; python_version >= "3.5"
pydantic==1.8.1
pyparsing==2.4.7
python-multipart==0.0.5
pytz==2021.1
pyyaml==5.4.1
requests==2.25.1
rsa==4.7.2; python_version >= "3.6"
six==1.15.0
starlette==0.13.6
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.4
uvicorn==0.13.4
